refactor(cef): carve window meta handlers into window/meta.rs
Fifth sub-PR of the commands/window.rs modularization (Plan 1). Pure
move — no behavior change.

Extracted to window/meta.rs (all pub, dispatched by ipc.rs, re-exported
`pub use meta::*`): get_zoom_factor, set_zoom_factor, get_window_label,
is_main_window, get_double_click_time, list_window_instances,
list_windows, focus_window, get_instance_number, register_backend_window,
toggle_devtools, inspect_element_at.

Self-contained — every cross-module reference is a fully-qualified
crate:: path (events / ui_tasks / client / launcher_ipc), no helper
imports. mod.rs now holds only the window-creation cluster
(dev_vite_port, resolve_frontend_base_url, open_new_window,
open_subwindow, ...) + FrontendUrlError, the subject of the final carve.

Verified: `cargo check -p agentmux-cef --release` clean.
