+++ .changesets/1780033671-refactor-cef-extract-window-meta-handlers-into-commands-wind-tl3a.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(cef): extract window meta handlers into commands/window/meta.rs
